Extract tile_set.py and cleanup
